analyst update functionality
