+++ backend/routes/api/analyst.js
@@ -121,15 +121,8 @@ router.get("/archive", (req, res) => {
 // @route GET api/analyst/update/:id
 // @description Update article
 // @access Public
-router.put("update/:id", (req, res) => {
-
-  Article.findByIdAndUpdate(req.params.id, {
-    analystClaims: req.body.analystClaims,
-    claimStrength: req.body.claimStrength,
-    evidence: req.body.evidence,
-    isForClaim: req.body.isForClaim,
-    method: req.body.method,
-  })
+router.put("/:id", (req, res) => {
+  Article.findByIdAndUpdate(req.params.id, req.body)
     .then((article) => {
       if (!article) {
         return res.status(404).json({ error: "Article not found" });
@@ -141,31 +134,6 @@ router.put("update/:id", (req, res) => {
     );
 });
 
-// @route GET api/analyst/:id/:claimStrength/:forClaim/:method/:evidence/:claim
-// @description Update article
-// @access Analyst
-router.put("/:id/:claimStrength/:forClaim/:method/:evidence/:claim", (req, res) => {
- 
-  Article.findByIdAndUpdate(
-    req.params.id, 
-    { strengthOfClaim: req.params.claimStrength,
-      isForClaim: req.params.forClaim,
-      method: req.params.method,
-      evidence: req.params.evidence,
-      analystClaims: req.params.claim,
-    },
-    { new: true } // Return the updated object
-  ) 
-  .then((article) => {
-    if (!article) {
-      return res.status(404).json({ error: "Article not found" });
-    }
-    return res.json({ msg: "Updated successfully", article });
-  })
-  .catch((err) =>
-    res.status(400).json({ error: "Unable to update the Database" })
-  );
-});
